backend/requirements.txt
django==3.0.1
djangorestframework==3.11.0
drf-yasg==1.17.0
djangorestframework-jwt==1.11.0
